+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-
+## Added
+- added mode JSON, LATTE and PLAIN_TEXT to CodeEditor
 
 ## [2.0.2] - 2023-03-29
 ### Fixed